+++ apps/restconf/restconf_lib.c
@@ -207,8 +207,8 @@ static const map_str2int http_media_map[] = {
     {"application/yang-data+json",       YANG_DATA_JSON},
     {"application/yang-patch+xml",       YANG_PATCH_XML},
     {"application/yang-patch+json",      YANG_PATCH_JSON},
-    {"application/yang.collection+xml",  YANG_COLLECTION_XML},
-    {"application/yang.collection+json", YANG_COLLECTION_JSON},
+    {"application/yang-collection+xml",  YANG_COLLECTION_XML},
+    {"application/yang-collection+json", YANG_COLLECTION_JSON},
     {NULL,                            -1}
 };